'help marti makecon packet'

+++ main.cpp
@@ -2,6 +2,7 @@
 #include "./Nodes/Node.h"
 #include "./Nodes/C.h"
 #include "./Nodes/T.h"
+#include "utils.h"
 
 Node* node;
 
@@ -13,13 +14,13 @@ int main(int argc, char* argv[]) {
 
     if (std::string(argv[1]) == "C") {
         try{
-            node = new C(std::string(argv[2]), atoi(argv[3]), std::string(argv[4]));
+            node = new C(std::string(argv[2]), (uint16_t) atoi(argv[3]), std::string(argv[4]));
         } catch (int e){
             std::cout << "Couldn't create C node";
         }
     } else if(std::string(argv[1]) == "T"){
         try{
-            node = new T(std::string(argv[2]), atoi(argv[3]), std::string(argv[4]));
+            node = new T(std::string(argv[2]), (uint16_t) atoi(argv[3]), std::string(argv[4]));
         } catch (int e){
             std::cout << "Couldn't create T node";
         }
